Clean up preprocessor
- move contents of cpp.rs into lib.rs
- move Definition/Definitions into definitions.rs
